Run integration tests in profile in parallel
