+++ test/integration/docker_test.go
@@ -27,7 +27,7 @@ import (
 )
 
 func TestDocker(t *testing.T) {
-	mk := NewMinikubeRunner(t, t.Name())
+	mk := NewMinikubeRunner(t, "")
 	if strings.Contains(mk.StartArgs, "--vm-driver=none") {
 		t.Skip("skipping test as none driver does not bundle docker")
 	}
