+++ test/integration/flags.go
@@ -18,7 +18,6 @@ package integration
 
 import (
 	"flag"
-	"fmt"
 	"os"
 	"testing"
 
@@ -38,15 +37,12 @@ var mountArgs = flag.String("minikube-mount-args", "", "Arguments to pass to min
 var testdataDir = flag.String("testdata-dir", "testdata", "the directory relative to test/integration where the testdata lives")
 
 // NewMinikubeRunner creates a new MinikubeRunner
-func NewMinikubeRunner(t *testing.T, profile string) util.MinikubeRunner {
-	if profile != "" {
-		a := fmt.Sprintf("%s -p %s", *args, profile)
-		args = &a
-	}
+func NewMinikubeRunner(t *testing.T, p string) util.MinikubeRunner {
 	return util.MinikubeRunner{
 		Args:       *args,
 		BinaryPath: *binaryPath,
 		StartArgs:  *startArgs,
+		Profile:    p,
 		MountArgs:  *mountArgs,
 		T:          t,
 	}
